解决循环依赖的问题
