+++ ui/categorybar.py
@@ -26,7 +26,7 @@ import sip
 from PyQt5.QtWidgets import *
 from PyQt5.QtGui import *
 from PyQt5.QtCore import *
-from models.enums import Signals
+from models.signals import Signals 
 
 
 class CategoryBar(QWidget,Signals):
